Finished L2 JWT and refresh token

+++ film-theater/Controllers/RoomsController.cs
@@ -7,6 +7,10 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
+using film_theater.Auth.Model;
+using Microsoft.AspNetCore.Identity;
+using film_theater.Data.Dtos.Auth;
 
 namespace film_theater.Controllers
 {
@@ -19,38 +23,43 @@ namespace film_theater.Controllers
        /api/theaters/{theaterId}/rooms/{roomId} DELETE 200/204     
     */
     [ApiController]
+    [Authorize(Roles = UserRoles.Moderator + "," + UserRoles.Admin)]
     [Route("api/theaters/{theaterId}/rooms")]
     public class RoomsController : ControllerBase
     {
         private readonly ITheatersRepository _theatersRepository;
         private readonly IRoomsRepository _roomsRepository;
         private readonly IMapper _mapper;
+        private readonly IAuthorizationService _authorizationService;
 
-        public RoomsController(ITheatersRepository theatersRepository, IRoomsRepository roomsRepository, IMapper mapper)
+        public RoomsController(ITheatersRepository theatersRepository,
+            IRoomsRepository roomsRepository, IMapper mapper,
+            IAuthorizationService authorizationService)
         {
             _theatersRepository = theatersRepository;
             _roomsRepository = roomsRepository;
             _mapper = mapper;
+            _authorizationService = authorizationService;
         }
 
         [HttpGet]
-        public async Task<IEnumerable<RoomDto>> GetAll(int theaterId)
+        public async Task<ActionResult<IEnumerable<RoomDto>>> GetAll(int theaterId)
         {
             var theaters = await _roomsRepository.GetAll(theaterId);
 
             // 200
-            return theaters.Select(o => _mapper.Map<RoomDto>(o));
+            return Ok(theaters.Select(o => _mapper.Map<RoomDto>(o)));
         }
 
         // /api/theaters/1/Rooms/2
         [HttpGet("{roomId}")]
         public async Task<ActionResult<RoomDto>> Get(int theaterId, int roomId)
         {
-            var theater = await _roomsRepository.Get(theaterId, roomId);
-            if (theater == null) return NotFound();
+            var room = await _roomsRepository.Get(theaterId, roomId);
+            if (room == null) return NotFound();
 
             // 200
-            return Ok(_mapper.Map<RoomDto>(theater));
+            return Ok(_mapper.Map<RoomDto>(room));
         }
 
         [HttpPost]
@@ -59,8 +68,13 @@ namespace film_theater.Controllers
             var theater = await _theatersRepository.Get(theaterId);
             if (theater == null) return NotFound();
 
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, theater, PolicyNames.SameUser);
+            if (!authorizationResult.Succeeded)
+                return Forbid();
+
             var room = _mapper.Map<Room>(roomDto);
             room.TheaterId = theaterId;
+            room.UserId = User.FindFirst(CustomClaims.UserId).Value;
 
             await _roomsRepository.Post(room);
 
@@ -71,18 +85,19 @@ namespace film_theater.Controllers
         [HttpPut("{roomId}")]
         public async Task<ActionResult<RoomDto>> Put(int theaterId, int roomId, UpdateRoomDto roomDto)
         {
-            var theater = await _theatersRepository.Get(theaterId);
-            if (theater == null) return NotFound();
+            var room = await _roomsRepository.Get(theaterId, roomId);
+            if (room == null) return NotFound();
 
-            var oldRoom = await _roomsRepository.Get(theaterId, roomId);
-            if (oldRoom == null) return NotFound();
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, room, PolicyNames.SameUser);
+            if (!authorizationResult.Succeeded)
+                return Forbid();
 
-            _mapper.Map(roomDto, oldRoom);
+            _mapper.Map(roomDto, room);
 
-            await _roomsRepository.Put(oldRoom);
+            await _roomsRepository.Put(room);
 
             // 200
-            return Ok(_mapper.Map<RoomDto>(oldRoom));
+            return Ok(_mapper.Map<RoomDto>(room));
         }
 
         [HttpDelete("{roomId}")]
@@ -90,6 +105,10 @@ namespace film_theater.Controllers
         {
             var room = await _roomsRepository.Get(theaterId, roomId);
             if (room == null) return NotFound();
+
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, room, PolicyNames.SameUser);
+            if (!authorizationResult.Succeeded)
+                return Forbid();
 
             await _roomsRepository.Delete(room);
